### 2023-09-11 #5837 @editorialbot
[PRE REVIEW]: ORMIR_XCT: A Python package for high resolution peripheral quantitative computed tomography image processing
**Submitting author:** @mkuczyns (<a href="http://orcid.org/0000-0001-5671-1205">Michael Tadeusz Kuczynski</a>)
**Repository:** https://github.com/SpectraCollab/ORMIR_XCT
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
