### 2023-09-19 e4b2c1e
Added DOIs to references and changed headers in paper
