+++ manuscript/paper.bib
@@ -1,5 +1,6 @@
 @article{Brunet:2020,
-  	url = {https://doi.org/10.1186/s12880-020-00437-8},
+  	url = {https://link.springer.com/article/10.1186/s12880-020-00437-8},
+	doi = {10.1186/s12880-020-00437-8},
   	Archiveprefix = {},
   	Author = {{Brunet}, S.C., {Kuczynski}, M.T., {Bhatla}, J.L., {Lemay}, S., {Pauchard}, Y., {Salat}, P., {Barnabe}, C., {Manske}, S.L.},
   	Eprint = {},
@@ -11,7 +12,8 @@
 }
 
 @article{Burt:2016,
-  	url = {https://doi.org/10.1002/jbmr.2873},
+  	url = {https://pubmed.ncbi.nlm.nih.gov/27192388/},
+	doi = {10.1002/jbmr.2873},
   	Archiveprefix = {},
   	Author = {{Burt}, L.A., {Liang}, Z., {Sajobi}, T.T., {Hanley}, D.A., {Boyd}, S.K.},
   	Eprint = {},
@@ -23,7 +25,8 @@
 }
 
 @article{Brunet:2021,
-  	url = {https://doi.org/10.1080/03009742.2020.1869303},
+  	url = {https://pubmed.ncbi.nlm.nih.gov/33775211/},
+	doi = {10.1080/03009742.2020.1869303},
   	Archiveprefix = {},
   	Author = {{Brunet}, S.C., {Tse}, J.J., {Kuczynski}, M.T., {Engelke}, K., {Boyd}, S.K., {Barnabe}, C., {Manske}, S.L.},
   	Eprint = {},
@@ -35,7 +38,8 @@
 }
 
 @article{Burghardt:2010,
-  	url = {https://doi.org/10.1016/j.bone.2010.05.034},
+  	url = {https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2926164/},
+	doi = {0.1016/j.bone.2010.05.034},
   	Archiveprefix = {},
   	Author = {{Burghardt}, A.J., {Buie}, H.R., {Laib}, A., {Majumdar}, S., {Boyd}, S.K.},
   	Eprint = {},
@@ -47,7 +51,8 @@
 }
 
 @article{Hildebrand:1996,
-  	url = {https://doi.org/10.1046/j.1365-2818.1997.1340694.x},
+  	url = {https://onlinelibrary.wiley.com/doi/abs/10.1046/j.1365-2818.1997.1340694.x},
+	doi = {10.1046/j.1365-2818.1997.1340694.x},
   	Archiveprefix = {},
   	Author = {{Hildebrand}, T. and {Ruegsegger}, P.},
   	Eprint = {},
@@ -59,7 +64,8 @@
 }
 
 @article{Harris:2020,
-  	url = {https://doi.org/10.1038/s41586-020-2649-2},
+  	url = {https://pubmed.ncbi.nlm.nih.gov/32939066/},
+	doi = {10.1038/s41586-020-2649-2},
   	Archiveprefix = {},
[…]
