+++ manuscript/paper.bib
@@ -39,7 +39,7 @@
 
 @article{Burghardt:2010,
   	url = {https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2926164/},
-	doi = {0.1016/j.bone.2010.05.034},
+	doi = {10.1016/j.bone.2010.05.034},
   	Archiveprefix = {},
   	Author = {{Burghardt}, A.J., {Buie}, H.R., {Laib}, A., {Majumdar}, S., {Boyd}, S.K.},
   	Eprint = {},
