### 2023-10-11 #5837 @spinicist
Apologies, I am too busy at the moment. Also I'm very surprised my paper/repo came top of the similarity list. It's an MRI project, not CT. But I guess the generic name might not help.
