### 2023-10-13 #5837 @prashjha
Hi @spinicist, no worries. Yes, the similarity check is the new feature JOSS introduced, and there is a scope for improvement.
