### 2023-10-13 #5837 @prashjha
Hi @po60nani, I am looking for one more reviewer for this JOSS submission. As per the JOSS similarity check, it appears this submission is relevant to your previous JOSS publication. You can read more about this work by following the pdf [link](https://github.com/openjournals/joss-papers/blob/joss.05837/joss.05837/10.21105.joss.05837.pdf) and software [link](https://github.com/SpectraCollab/ORMIR_XCT).

Please let me know if you will be interested in reviewing this work. Thanks!!
