### 2023-11-23 #5837 @prashjha
Hi @ajinkyakadu, would you be interested in reviewing [this JOSS submission](https://github.com/openjournals/joss-papers/blob/joss.05837/joss.05837/10.21105.joss.05837.pdf)? The corresponding software is at this [link](https://github.com/SpectraCollab/ORMIR_XCT). This [link](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) covers pretty much everything about the review process; I will be happy to help with more information.

Please let me know if you are interested. Thank you!
