### 2023-12-23 #6084 @zhangjy-ge
Hi,

I have finished my preliminary review of the manuscript and I think it satisfies most requirements, however there are also something that may be improved:

1. unit tests are not all done and some are still unimplemented.
2. I did not find any manual or documentations about how to use the package.
3. There are no clear guidelines about how to contribute/report bugs
4. In the installation step, maybe add a first step of something like _download the repository_ will make everything clearer.
