### 2024-01-05 #6084 @codeling
I have now finished my first pass as well.
I second the points made by @zhangjy-ge, and have a few additional remarks:

- The datasets
- When installing as mentioned via `conda env create -f environment.yml`, I got the error
```
ERROR: Could not find a version that satisfies the requirement itk-numerics==5.2.1 (from versions: 5.0.1, 5.1.0, 5.1.0.)ERROR: No matching distribution found for itk-numerics==5.2.1
```
I'm not too familiar with conda/pip details, and find the error message bit confusing. If I activate the environment (it is created despite the error) and check installed packages (`pip list` / `conda list`), I find that itk packages (itk, itk-core, itk-filtering, itk-numerics, itk-registration) are installed in version 5.3.0 already (from channel pypi - they seem to be pulled in from somewhere else, possibly as simpleitk dependency?). If I change all ITK packages from "5.2.1" to "5.3.0", the environment creation proceeds without errors.

- Step 2 is supposed to be run inside the created environment, right? I would recommend adding a `conda activate ormir_xct` step before

- at this point I cannot comment about whether the above produces a working environment, since I haven't found a documentation on how to use
