### 2024-02-12 f2ef4b1
Updated README to include activating conda env
