### 2024-02-14 #6084 @mkuczyns
@prashjha @Kevin-Mattheus-Moerman: Sorry for the delay in responding to the reviewer's comments. We have made some changes to address the unclear installation instructions and added a few more example Jupyter Notebooks to demonstrate the package's functionality. We are currently working on completing the missing unit tests and creating Sphinx documentation.

As for data availability, we have uploaded the Python-generated shapes to Zenodo (see paper section on thickness calculation). The HR-pQCT image data that was used for the other sections is taken from studies that are currently awaiting publication and can be made available once those studies are published.
