### 2024-04-05 #6084 @codeling
The now provided example notebooks are great as introduction and documentation.
There are very minor issues that I came across while testing, which I reported as separate issues in the repository, but these are definitely not showstoppers for going ahead here, from my side the paper/package can be accepted!
