+++ environment.yml
@@ -7,6 +7,7 @@ channels:
 dependencies:
   - itkwidgets=0.32.0
   - numpy=1.23.3
+  - pandas=2.0.3
   - pbr=5.4.2
   - pip=21.2.4
   - python=3.8.5
@@ -16,14 +17,14 @@ dependencies:
   - simpleitk=2.0.2
   - vtk=9.0.1
   - pip:
-    - itk==5.2.1
-    - itk-core==5.2.1
-    - itk-filtering==5.2.1
-    - itk-io==5.2.1
+    - itk==5.3.0
+    - itk-core==5.3.0
+    - itk-filtering==5.3.0
+    - itk-io==5.3.0
     - itk-ioscanco==0.9.2
-    - itk-numerics==5.2.1
-    - itk-registration==5.2.1
-    - itk-segmentation==5.2.1
+    - itk-numerics==5.3.0
+    - itk-registration==5.3.0
+    - itk-segmentation==5.3.0
     - notebook==6.5.2
     - numba==0.56.3
     - kimimaro==3.3.0
