### 2024-05-03 #6084 @mkuczyns
Hi @prashjha, I am just wondering if there is an update regarding the acceptance of this publication to JOSS? Thanks!
