### 2024-05-10 #6084 @prashjha
Hi @mkuczyns, see if these suggestions make sense for your draft:
- Section and sub-section headings are not standard, e.g., numbering is not needed, dash between number and text and colon after text are not desired. See JOSS example paper format [here](https://joss.readthedocs.io/en/latest/example_paper.html)
- In section "Autocontour" and Figure 1, it will be useful to mention that the value closer to 1 for DICE and Jaccard columns is desired meaning the new implementation is accurate.
- In the caption of Figure 1, You mention both Figure 1 and Table 1. Maybe remove "Table 1" and just say "DICE coefficients, Jaccard indices, and Hausdorff distances (voxels) between IPL and ORMIR_XCT automatic contouring implementations for three type of datasets." Similar issue is found in the caption of Figure 5.
- The caption of Figure 2 starts with "Figure 1:". Same issue with captions of Figure 3 and 4.
