### 2024-05-13 f164def
Fixes for final JOSS submission
