+++ manuscript/paper.md
@@ -69,44 +69,44 @@ date: 30 August 2023
 bibliography: paper.bib
 ---
 
-# 1 – Statement of need:
+# Statement of need:
 High resolution peripheral quantitative computed tomography (HR-pQCT) is an imaging technique with a nominal isotropic voxel size of 61µm capable of imaging trabecular bone *in-vivo*. HR-pQCT has a wide range of applications, primarily focused on bone to improve our understanding of musculoskeletal diseases [@Brunet:2020], assess epidemiological associations  [@Burt:2016], and evaluate the effects of pharmaceutical interventions [@Brunet:2021]. Processing HR-pQCT images has largely been supported using the scanner manufacturer’s scripting language (Image Processing Language, IPL, Scanco Medical). However, by expanding image processing workflows outside of the scanner manufacturer’s software environment, users have the flexibility to apply more advanced mathematical techniques and leverage modern software packages to improve image processing. The `ORMIR_XCT` Python package was developed to reimplement some existing IPL workflows and provide an open and reproducible package allowing for the development of advanced HR-pQCT data processing workflows.
 
 
-# 2 – Package summary:
+# Package summary:
 The development of this package began during the Jupyter Community Workshop in Maastricht, Netherlands in June 2022, hosted by the Open and Reproducible Musculoskeletal Imaging Research (ORMIR) group. During this workshop, the conceptualization and initial development of the `ORMIR_XCT` package began with support from Kitware. The `ORMIR_XCT` package is currently being used by members of the ORMIR community to assess bone mineral density (BMD) and joint space width (JSW) changes in HR-pQCT scans of patients with osteoarthritis or rheumatoid arthritis.
 
 The `ORMIR_XCT` package contains four modules for processing HR-pQCT data of bones and joints: 1) automatic contouring of the periosteal surface, 2) JSW analysis, 3) BMD calculation, and 4) segmentation of trabecular bone. Analyses have been performed to compare outputs from the `ORMIR_XCT` package to results generated using IPL. Jupyter Notebook examples are provided to describe the workflows implemented in the `ORMIR_XCT` package.
 
-# 3 – Comparison to IPL: 
-## 3.1 - Autocontour:
-Automatic periosteal contouring (autocontour) was performed on a sample dataset of HR-pQCT images of the 2nd and 3rd distal interphalangeal (DIP) and trapeziometacarpal (TMC) joints (n = 59). The images came from 10 participants in an osteoarthritic group and 10 participants in a control group. This dataset is representative of typical data that is collected using HR-pQCT. Binarized images were obtained using both the IPL standard method for automatic periosteal contouring [@Burghardt:2010] and the `ORMIR_XCT` package. Segmentation were compared by computing DICE coefficients, Jaccard indices, and Hausdorff distances between images. The IPL method for automatic periosteal contouring was taken as the ground truth. Results of the segmentation comparison are shown in \autoref{fig:Table1}.
+# Comparison to IPL: 
+## Autocontour:
[…]
