### 2024-05-13 #6084 @mkuczyns
Hi @prashjha,

Thank you for your comments. I have made changes to the manuscript as per your suggestions.
- A Zenodo archive has been created at the following DOI: [https://doi.org/10.5281/zenodo.11186589](https://doi.org/10.5281/zenodo.11186589)
- The code has been updated since the submission of this package.
- Numbers and dashes have been removed from section and sub-section headings.
- The "Autocontour" section and Figure 1 caption have been updated to indicate the range for DICE coefficients and Jaccard indices.
- The figure captions have been fixed.

Let me know if there are any other changes needed!
