### 2024-05-14 #6084 @prashjha
@mkuczyns, thank you. I only see v1.0.0 in the release, which is already attached to the JOSS submission.
