### 2024-05-14 #6084 @mkuczyns
@prashjha: I have removed the colons from all headings and subheadings. Sorry for the confusion with the version number. The current version number attached to the JOSS submission has the update to date code.

Let me know if there is anything else!
