### 2024-05-25 #6084 @Kevin-Mattheus-Moerman
- [ ] Please edit the ZENODO archive listed licenses (see Rights) to match your software license.
- [ ] The text mixes European English (e.g. `Analyses`, and `neighbouring`) and American English e.g. `Binarized`, `optimized, `skeletonization`. Perhaps choose one convention and stick to it consistently.
- [ ] In your affiliations, please spell out `USA` as `United States of America`.
- [ ] You paper currently lacks an "Acknowledgements" section (e.g. of funding etc). If this was not your intention you may add such a section now.
