### 2024-05-29 #6084 @mkuczyns
Hi @Kevin-Mattheus-Moerman,

Sorry for the delay, I was getting funding information from some of the co-authors. I have made the following changes as per your previous comment:

- The Zenodo license has been updated to match the license on the GitHub page.
- The manuscript has been updated to not mix American and UK English.
- The manuscript has been updated to spell out United States of America.
- An acknowledgement section has been added to the manuscript.

Please let me know if there is anything else!
