### 2024-05-30 #6084 @Kevin-Mattheus-Moerman
@mkuczyns thanks for making those changes. One more point which I initially missed: please include the `v` in `v1.0.0` for the ZENODO listed archive. Can you add this please?
