### 2024-05-30 #6084 @mkuczyns
@Kevin-Mattheus-Moerman: I updated the Zenodo version to v1.0.0!
